test: ensure core and CLI token parsing share map
Intent: consolidate token parsing map into shared module
Operators involved: [Transition, Recursivity]
Affected invariants: [#4, #7]
Key changes:
- Introduce shared TOKEN_MAP and THOL parser module
- Add equivalence test between core and CLI parsers
Expected risks/dissonances: shared map mutations require caution
Metrics: (C(t), Si, νf, phase) stable
Equivalence map: TOKEN_MAP -> tnfr.token_map.TOKEN_MAP
